Store ext for later use.
My last request had a bug in it. I added this.ext and used it for selective logging. Then I accidentally removed it from the official pull request. I later use this.ext to determine whether to do coffee script compilation. I've checked this.meta and ext isn't directly available in it.