Support coffee script in assemblies.
Detect a file has coffee script in it by checking that it's path ends with ".coffee"
Compile the code without a surrounding function so that all of the symbols are visible to the rest of the module. This is accomplished with the "bare" option to compile.
Add a header to the code to help reinforce that the code in the file is generated from coffee script and to make it obvious what version of coffee script was used.