I added a few features that the new Angular version of Family Tree needs to simplify our coding.
The first allows a single assembly to include sub assemblies into one file. This can greatly reduce the number of JavaScript files that you need to include in the assets section of your EJS file.

The second allows a developer to create a folder 'templates' and add a series of HTML files that each are made into different templates. This allows the developers to separate out different sections of templates into separate files and, thus, separate variables in their environment.

The templates are accessible using the templateList variable and the member variables are the names of the html files. (Minus the .html)

For example, if you have the following directory structure:
templates/mine.html
templates/yours.html
templates/testing.html

Then you can access those templates by using templateList.mine or templateList.yours and templateList.testing.

# Problem space:
- The new Angular tree needed a way to reduce include files.
- The new Angular tree needed to have several smaller templates instead of just one large template.
- When an assembler file uses 'simpleWrap': true, and is loaded by inject then window gets messed up.

# Changes:
- Include all templates located in the templates folder. Store them as members of the templateList variable. If no templates folder then no templates are added.
- Include all sub-assemblies if they are defined
- Corrected the IIFE function to pass 'window' and not 'this' since inject messes up 'this'

# Tested in languages:
- [X] DE
- [X] EN
- [X] ES
- [X] FR
- [X] IT
- [X] JA
- [X] KO
- [X] PT
- [X] RU
- [X] ZH